% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Convertir grados Celsius a Fahrenheit}
\usage{
celsius_to_fahrenheit(celsius)
}
\arguments{
\item{celsius}{Un valor numerico o vector numerico que representa grados Celsius.}
}
\value{
Un valor numerico o vector numerico en grados Fahrenheit.
}
\description{
Convierte una temperatura o vector de temperaturas de grados Celsius a Fahrenheit.
}
\examples{
celsius_to_fahrenheit(0)
celsius_to_fahrenheit(c(10, 25, 30))
}
