% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos}
\alias{datos}
\title{Conjunto de datos meteorológicos}
\format{
Un data frame con 2 filas y 4 columnas:
\describe{
\item{estacion}{Nombre de la estación meteorológica (carácter).}
\item{temperatura}{Temperatura media diaria en grados Celsius (numérico).}
\item{humedad}{Humedad relativa promedio en porcentaje (numérico).}
\item{precipitacion}{Precipitación diaria en milímetros (numérico).}
}
}
\source{
Datos simulados para pruebas.
}
\usage{
datos
}
\description{
Este conjunto de datos incluye información simulada sobre estaciones meteorológicas.
}
\keyword{datasets}
