% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Descargar datos meteorológicos}
\usage{
descargar_datos(id_estacion, directorio_destino = tempdir())
}
\arguments{
\item{id_estacion}{ID de la estación meteorológica. Los posibles son: NH0098, NH0046, NH437, NH472 y NH0910.}

\item{directorio_destino}{Directorio donde se guardará el archivo CSV. Por defecto, se usa un directorio temporal.}
}
\value{
Un data.frame con los datos meteorológicos de la estación.
Además, se guarda un archivo CSV en el directorio especificado.
}
\description{
La función \code{descargar_datos} toma el ID de una estación y descarga un archivo CSV con los datos meteorológicos.
Por defecto, se guardará en un directorio temporal seguro, pero se puede especificar otro destino.
}
\examples{
# Descargar datos de la estación NH0098 en un directorio temporal:
datos <- descargar_datos("NH0098")

# O especificar un directorio personalizado:
# datos <- descargar_datos("NH0098", "mi_carpeta")
}
