% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar_temperatura_mensual.R
\name{graficar_temperatura_mensual}
\alias{graficar_temperatura_mensual}
\title{Gráfico de Temperatura Promedio Mensual}
\usage{
graficar_temperatura_mensual(
  datos,
  colores = c("green", "blue", "red", "yellow", "purple"),
  titulo = "Temperatura Mensual"
)
}
\arguments{
\item{datos}{tibble o dataset al cual se le quiere realizar el gráfico de la temperatura promedio mensual.}

\item{colores}{colores con los que se quiere que aparezca la temperatura en el gráfico de linea.}

\item{titulo}{frase o palabra que se desea poner de titulo en el gráfico.}
}
\value{
imagen de un gráfico de lineas.
}
\description{
La función graficar_temperatura_mensual sirve para graficar el promedio de la temperatura mensual del dataset proporcionado
}
\examples{
graficar_temperatura_mensual(estaciones_unidas,c("green", "blue", "red","yellow","purple"),"Temperatura Promedio Mensual Estación 1")
graficar_temperatura_mensual(estaciones_unidas,c("purple", "pink", "lightblue","orange","brown"),"Temperatura Promedio Mensual Estación de Argentina-Rosario")

}
