% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_estaciones.R
\name{leer_estaciones}
\alias{leer_estaciones}
\title{Lee archivos sobre estaciones meteorológicas}
\usage{
leer_estaciones(id_estacion, ruta)
}
\arguments{
\item{id_estacion}{ID de la estación cuyos datos se quieren leer}

\item{ruta}{Ruta o directorio en donde se descargarán y leeran los datos}
}
\value{
Los datos en forma de .csv para poder almacenarlos en una variable y utilizarlos
}
\description{
La funcion leer_estaciones sirve para descargar y leer los datos de estaciones a partir de su ID. Tambien debe proporcionarse la ruta en donde se descargara el archivo de los datos de la estacion
}
\examples{
leer_estaciones("NH0437","datos/NH0437.csv")
leer_estaciones("NH0910","datos/NH0910.csv")

}
