% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos números}
\usage{
suma(x, y)
}
\arguments{
\item{x}{El primer número (número).}

\item{y}{El segundo número (número).}
}
\value{
La suma de los dos números (número).
}
\description{
Esta función toma dos números y devuelve su suma.
}
\examples{
suma(2, 3) # Devuelve 5
suma(-1, 5) # Devuelve 4
}
