% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Resumen Estadístico de Temperaturas por Estación}
\usage{
tabla_resumen_temperatura(estacion)
}
\arguments{
\item{estacion}{El conjunto de datos de estaciones
meteorológicas o de una estacion en especifico. Se utilizaran las columnas:
\itemize{
\item \code{id}: Identificador de la estacion.
\item \code{temperatura_abrigo_150cm}: Columna con los valores de temperatura
utilizados para los cálculos.
}}
}
\value{
Un objeto tibble que contiene una fila por
cada estación (\code{id}) con las siguientes columnas de resumen:
\code{id}, \code{Max}, \code{Min}, \code{Promedio}, y \code{Desvio_Estandar}.
}
\description{
Calcula un resumen estadístico clave para la columna de temperatura
(\code{temperatura_abrigo_150cm}) en el conjunto de datos de una o varias estaciones.
}
\details{
La función agrupa los datos por estación (\code{id}) y devuelve una tabla con la
temperatura máxima, mínima, promedio y desviación estándar. Es útil para
obtener una visión general del comportamiento de la temperatura en cada estación.

Esta función requiere el paquete \code{dplyr} para la manipulación de datos.
}
\examples{
\dontshow{
ruta_temp1 <- file.path(tempdir(), "NH0437.csv")
ruta_temp2 <- file.path(tempdir(), "NH0910.csv")

NH0437 <- leer_estaciones("NH0437", ruta_temp1)
NH0910 <- leer_estaciones("NH0910", ruta_temp2)
}

tabla_resumen_temperatura(NH0437)
tabla_resumen_temperatura(NH0910)

}
