% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Descargar datos meteorologicos}
\usage{
descargar_datos(id_estacion, directorio_destino = tempdir())
}
\arguments{
\item{id_estacion}{ID de la estacion (ej.: NH0098, NH0046, NH0437, NH0472, NH0910).}

\item{directorio_destino}{Carpeta donde guardar el CSV. Por defecto usa un tempdir seguro.}
}
\value{
Un data.frame con los datos descargados.
}
\description{
Descarga el CSV de una estacion y lo lee a un data.frame.
}
