% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_estaciones.R
\name{leer_estaciones}
\alias{leer_estaciones}
\title{Descarga y lee los datos de archivos de estaciones meteorológicas}
\usage{
leer_estaciones(id_estacion, ruta)
}
\arguments{
\item{id_estacion}{ID de la estación cuyos datos se quieren leer.}

\item{ruta}{Ruta o directorio en donde se descargarán y leerán los datos}
}
\value{
Los datos en formato CSV para poder almacenarlos en una variable y utilizarlos posteriormente.
}
\description{
Descarga un archivo formato CSV con  datos de una estación meteorológica
específica y lo guarda en una ruta local. Si el archivo ya existe en la ruta especificada, la función lo lee directamente
sin volver a descargarlo.
}
\details{
Requiere los paquetes \code{glue}, \code{readr} y \code{cli}.
}
\examples{
\dontshow{
ruta_temp1 <- file.path(tempdir(), "NH0437.csv")
ruta_temp2 <- file.path(tempdir(), "NH0910.csv")
}
leer_estaciones("NH0437", ruta_temp1)
leer_estaciones("NH0910", ruta_temp2)

}
