% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_estaciones.R
\name{leer_estaciones}
\alias{leer_estaciones}
\title{Descarga y lectura de datos meteorológicos de una estación}
\usage{
leer_estaciones(id_estacion, ruta)
}
\arguments{
\item{id_estacion}{ID de la estación cuyos datos se quieren leer.}

\item{ruta}{Ruta o directorio en donde se descargarán y leerán los datos}
}
\value{
Los datos en formato CSV para poder almacenarlos en una variable y utilizarlos posteriormente.
}
\description{
Esta función descarga (si es necesario) y lee los datos de una estación meteorológica
específica en formato CSV. Si el archivo ya existe en la ruta indicada, se utiliza el
archivo local sin volver a descargarlo.
}
\details{
Requiere los paquetes \code{glue}, \code{readr} y \code{cli}.
}
\examples{
\dontshow{
ruta_temp1 <- file.path(tempdir(), "NH0437.csv")
ruta_temp2 <- file.path(tempdir(), "NH0910.csv")
}
leer_estaciones("NH0437", ruta_temp1)
leer_estaciones("NH0910", ruta_temp2)

}
