% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Tabla con un resumen numérico de las temperaturas de una estación}
\usage{
tabla_resumen_temperatura(estacion)
}
\arguments{
\item{estacion}{tibble o data.frame con los datos a lo que se les va a realizar el resumen numérico}
}
\value{
tibble o data.frame con el mínimo, máximo, promedio y desvio estandar de la temperatura
}
\description{
La función tabla_resumen_temperatura devuelve un resumen numérico de las temperaturas de una estacion, la cual se pasa como parámetro. Esta tabla contiene: temperatura máxima, temperatura mínima, temperatura promedio y su desvio estandar.
}
\examples{
tabla_resumen_temperatura(estacion_1) # estacion_1 <- read.csv("datos/estacion_1.csv")
tabla_resumen_temperatura(estacion_2) # estacion_2 <- read.csv("datos/estacion_2.csv")

}
