% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Resumen Estadístico de Temperaturas por Estación}
\usage{
tabla_resumen_temperatura(estacion)
}
\arguments{
\item{estacion}{El conjunto de datos de estaciones
meteorológicas o de una estacion en especifico. Se utilizaran las columnas:
\itemize{
\item \code{id}: Identificador de la estacion.
\item \code{temperatura_abrigo_150cm}: Columna con los valores de temperatura
utilizados para los cálculos.
}}
}
\value{
Un objeto tibble que contiene una fila por
cada estación (\code{id}) con las siguientes columnas de resumen:
\code{id}, \code{Max}, \code{Min}, \code{Promedio}, y \code{Desvio_Estandar}.
}
\description{
Calcula un resumen estadistico clave para la columna de temperatura
(\code{temperatura_abrigo_150cm}) en el conjunto de datos de una o varias estaciones.

La tabla de salida incluye la temperatura máxima, mínima, promedio y la
desviación estándar. Los cálculos se agrupan por el identificador de la
estación (\code{id}).

Esta función requiere el paquete \code{dplyr} para la manipulación de datos.
}
\examples{
tabla_resumen_temperatura(estacion_1)
tabla_resumen_temperatura(datos_combinados)

}
